+++ packages/subgraph/tests/plugin/pluginRepo.test.ts
@@ -2,59 +2,33 @@ import {assert, clearStore, test} from 'matchstick-as/assembly/index';
 import {ADDRESS_ONE, ADDRESS_TWO} from '../constants';
 import {createVersionCreated} from './utils';
 import {handleVersionCreated} from '../../src/plugin/pluginRepo';
-import {Address} from '@graphprotocol/graph-ts';
+import {Address, Bytes} from '@graphprotocol/graph-ts';
 
-// TODO: SARKAWT
-// test('versionCreated event', () => {
-//   let event = createVersionCreated('0', ['1', '0', '0'], ADDRESS_ONE, '0x42');
+test('versionCreated event', () => {
+  let event = createVersionCreated(
+    '1',
+    '1',
+    ADDRESS_ONE,
+    Bytes.fromHexString('0x1234')
+  );
+  handleVersionCreated(event);
+  let id = `${event.address.toHexString()}_1_1`;
 
-//   handleVersionCreated(event);
+  assert.entityCount('PluginVersion', 1);
+  assert.fieldEquals('PluginVersion', id, 'id', id);
+  assert.fieldEquals(
+    'PluginVersion',
+    id,
+    'pluginRepo',
+    event.address.toHexString()
+  );
+  assert.fieldEquals('PluginVersion', id, 'release', '1');
+  assert.fieldEquals('PluginVersion', id, 'build', '1');
+  assert.fieldEquals('PluginVersion', id, 'pluginSetup', ADDRESS_ONE);
+  assert.fieldEquals('PluginVersion', id, 'metadata', '0x1234');
 
-//   let entityId = `${event.address.toHexString()}_0`;
-//   assert.fieldEquals(
-//     'PluginVersion',
-//     entityId,
-//     'pluginRepo',
-//     event.address.toHexString()
-//   );
-//   assert.fieldEquals(
-//     'PluginVersion',
-//     entityId,
-//     'semanticVersion',
-//     '[1, 0, 0]'
-//   );
-//   assert.fieldEquals(
-//     'PluginVersion',
-//     entityId,
-//     'pluginSetup',
-//     Address.fromString(ADDRESS_ONE).toHexString()
-//   );
-//   assert.fieldEquals('PluginVersion', entityId, 'contentURI', '0x42');
+  assert.entityCount('PluginSetup', 1);
+  assert.fieldEquals('PluginSetup', ADDRESS_ONE, 'id', ADDRESS_ONE);
 
-//   event = createVersionCreated('1', ['1', '1', '0'], ADDRESS_TWO, '0x43');
-
-//   entityId = `${event.address.toHexString()}_1`;
-//   handleVersionCreated(event);
-
-//   assert.fieldEquals(
-//     'PluginVersion',
-//     entityId,
-//     'pluginRepo',
-//     event.address.toHexString()
-//   );
-//   assert.fieldEquals(
-//     'PluginVersion',
-//     entityId,
-//     'semanticVersion',
-//     '[1, 1, 0]'
-//   );
-//   assert.fieldEquals(
-//     'PluginVersion',
-//     entityId,
-//     'pluginSetup',
-//     Address.fromString(ADDRESS_TWO).toHexString()
-//   );
-//   assert.fieldEquals('PluginVersion', entityId, 'contentURI', '0x43');
-
-//   clearStore();
-// });
+  clearStore();
+});